--- core/shared/src/main/scala/fs2/async/immutable/immutable.scala
@@ -1,4 +1,0 @@
-package fs2.async
-
-/** Provides types which allow asynchronous reading (but not writing). */
-package object immutable
