--- core/shared/src/main/scala/fs2/async/mutable/mutable.scala
@@ -1,4 +1,0 @@
-package fs2.async
-
-/** Provides types which allow asynchronous reading and writing. */
-package object mutable
