+++ core/shared/src/test/scala/fs2/async/OnceSpec.scala
@@ -8,15 +8,17 @@ import cats.implicits._
 import scala.concurrent.duration._
 import org.scalatest.EitherValues
 
+import fs2.concurrent.once
+
 class OnceSpec extends AsyncFs2Spec with EitherValues {
 
-  "async.once" - {
+  "once" - {
 
     "effect is not evaluated if the inner `F[A]` isn't bound" in {
       val t = for {
         ref <- Ref.of[IO, Int](42)
         act = ref.update(_ + 1)
-        _ <- async.once(act)
+        _ <- once(act)
         _ <- timerIO.sleep(100.millis)
         v <- ref.get
       } yield v
@@ -30,7 +32,7 @@ class OnceSpec extends AsyncFs2Spec with EitherValues {
           val ns = s + 1
           ns -> ns
         }
-        memoized <- async.once(act)
+        memoized <- once(act)
         x <- memoized
         y <- memoized
         v <- ref.get
@@ -45,7 +47,7 @@ class OnceSpec extends AsyncFs2Spec with EitherValues {
           val ns = s + 1
           ns -> ns
         }
-        memoized <- async.once(act)
+        memoized <- once(act)
         _ <- memoized.start
         x <- memoized
         _ <- timerIO.sleep(100.millis)
@@ -62,7 +64,7 @@ class OnceSpec extends AsyncFs2Spec with EitherValues {
           val ns = s + 1
           ns -> ns
         }
-        act2 = async.once(act1).flatten
+        act2 = once(act1).flatten
         _ <- Stream.repeatEval(act1).take(n).compile.drain.start
         _ <- Stream.repeatEval(act2).take(n).compile.drain.start
         _ <- timerIO.sleep(200.millis)
