+++ docs/migration-guide-1.0.md
@@ -21,7 +21,7 @@ The new `cats.effect.Timer` type was introduced in cats-effect 0.10. This type p
 
 ### Cancelation
 
-The `cats.effect.Concurrent` type class was introduced in cats-effect 0.10, providing the ability to start a `F[A]` computation as a lightweight thread and then either wait for the result or cancel the computation. This functionality is used throughout `fs2.async` to support cancelation of asynchronous tasks. Consider the use case of dequeuing an element from a queue and timing out if no element has been received after some specified duration. In FS2 0.10, this had to be done with `q.timedDequeue1`, as simply calling `dequeue1` and racing it with a timeout would leave some residual state inside the queue indicating there's a listener for data. FS2 0.10 had a number of similar methods throughout the API -- `timedGet`, `cancellableDequeue1`, etc. With cats-effect's new `Concurrent` support, these APIs are no longer needed, as we can implement cancelation in a composable fashion.
+The `cats.effect.Concurrent` type class was introduced in cats-effect 0.10, providing the ability to start a `F[A]` computation as a lightweight thread and then either wait for the result or cancel the computation. This functionality is used throughout `fs2.concurrent` (formerly `fs2.async`) to support cancelation of asynchronous tasks. Consider the use case of dequeuing an element from a queue and timing out if no element has been received after some specified duration. In FS2 0.10, this had to be done with `q.timedDequeue1`, as simply calling `dequeue1` and racing it with a timeout would leave some residual state inside the queue indicating there's a listener for data. FS2 0.10 had a number of similar methods throughout the API -- `timedGet`, `cancellableDequeue1`, etc. With cats-effect's new `Concurrent` support, these APIs are no longer needed, as we can implement cancelation in a composable fashion.
 
 A good example of the simplification here is the `fs2.async.Promise` type (now `cats.effect.concurrent.Deferred`, more on that later). In FS2 1.0, `Promise` has only 2 methods -- `get` and `complete`. Timed gets and cancelable gets can both be implemented in a straightforward way by combining `p.get` with `Concurrent[F].race` or `Concurrent[F].start`.
 
@@ -33,9 +33,9 @@ An exception to this change is the `fs2-io` module -- places where there's an in
 
 Another exception appears in the `fs2-io` module -- places where blocking calls are made to Java APIs (e.g., writing to a `java.io.OutputStream`). In such cases, an explicit blocking `ExecutionContext` must be passed. The blocking calls will be executed on the supplied `ExecutionContext` and then shifted back to the main asynchronous execution mechanism of the effect type (via `Timer[F].shift`).
 
-### Async Data Types
+### Concurrent Data Types
 
-Some of the data types from the `fs2.async` package have moved to `cats.effect.concurrent` -- specifically, `Ref`, `Promise` (now called `Deferred`), and `Semaphore`. As part of moving these data types, their APIs evolved a bit.
+Some of the data types from the old `fs2.async` package have moved to `cats.effect.concurrent` -- specifically, `Ref`, `Promise` (now called `Deferred`), and `Semaphore`. As part of moving these data types, their APIs evolved a bit.
 
 #### Ref
 
@@ -71,6 +71,31 @@ Some of the data types from the `fs2.async` package have moved to `cats.effect.c
 |`s.tryDecrementBy(n)`|`s.tryAcquireN(n)`|
 |`s.increment`|`s.release`|
 |`s.increment`|`s.release`|
+
+#### Signal
+
+The `fs2.async.immutable.Signal` type is now `fs2.concurrent.ReadableSignal` while `fs2.async.mutable.Signal` is replaced by either `fs2.concurrent.WritableSignal`, if you only need the ability to update the signal, or `fs2.concurrent.Signal`, if you need both read and write ability. Constructing a signal is now accomplished via `Signal.apply` instead of `fs2.async.signalOf`.
+
+#### Queue
+
+`Queue` also moved from `fs2.async.mutable.Queue` to `fs2.concurrent.Queue`. `Queue` now extends both `Enqueue` and `Dequeue`, allowing you to better delineate whether a function produces or consumes elements. Size information has been moved to `InspectableQueue`, so the runtime cost of maintaining size information isn't paid for all usages. Constructors are on the `Queue` and `InspectableQueue` companions -- e.g., `Queue.bounded(n)` or `Queue.synchronous`.
+
+#### Topic
+
+`Topic` moved from `fs2.async.mutable.Topic` to `fs2.concurrent.Topic` and the constructor has moved to `Topic.apply`.
+
+#### fs2.async Package Object
+
+The `fs2.async` package object contained constructors for concurrent data types and miscellaneous concurrency utilities (e.g., `start`, `fork`, `parallelTraverse`). Most of the miscellaneous concurrency utilities are no longer necessary because they are directly supported by cats. For example, `start` now exists on the `cats.effect.Concurrent` type class and `parTraverse` is available for any `Concurrent[F]`.
+
+One exception is `unsafeRunAsync`, which was removed from fs2 without a direct replacement in cats-effect. To run a computation asynchronously, you can use the following:
+
+```scala
+// Given F: ConcurrentEffect[F] & import cats.implicits._, cats.effect.implicits._
+fa.start.flatMap(_.join).runAsync(_ => IO.unit).unsafeRunSync
+```
+
+The only remaining concurrency utility is `fs2.concurrent.once`, which supports memoization of a task.
 
 ### Chunks and Segments
